VKCI-299: Refactors to allow CSI to use multiAZ implementation
