+++ pkg/vcdsdk/zonemap.go
@@ -1,4 +1,4 @@
-package config
+package vcdsdk
 
 import (
 	"fmt"
